Initial Commit
This is the inital commit for the game. The engine is almost entirely written. At this point I need to develop a GUI, and I'm still up in the air on how I'd like to go about doing that. I might go with JavaFX or perhaps turn this into a web-app. We'll see.